// Hackathon.Garbage.Web/ClientApp/src/app/models/google-maps-types.ts
export interface LatLngLiteral {
  lat: number;
  lng: number;
}